# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FamilySpace API - private family communication and experiences platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /planning-tasks:
    get:
      operationId: listPlanningTasks
      tags: [planning-tasks]
      summary: List planning tasks
      parameters:
        - name: calendarEventId
          in: query
          schema:
            type: string
        - name: invitationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of planning tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlanningTask"
  /planning-tasks/{id}:
    patch:
      operationId: updatePlanningTask
      tags: [planning-tasks]
      summary: Update a planning task (toggle complete, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanningTaskUpdate"
      responses:
        "200":
          description: Updated planning task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanningTask"
...
components:
  schemas:
...
    PlanningTask:
      type: object
      required: [id, title, category, completed, dueDate, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          enum: [childcare, reservation, tickets, transport, reminder, weather, backup, other]
        completed:
          type: boolean
        dueDate:
          type: ["string", "null"]
        calendarEventId:
          type: ["string", "null"]
        invitationId:
          type: ["string", "null"]
        assigneeId:
          type: ["string", "null"]
        createdAt:
          type: string
    PlanningTaskUpdate:
      type: object
      properties:
        completed:
          type: boolean
        title:
          type: string
        dueDate:
          type: string
        assigneeId:
          type: string
        description:
          type: string
...